+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Dependencies
 - Updated Mixin Extras version to 0.2.0-beta.9.
+- [FABRIC] Fixed incompatibility with zoom mods.
 - [FABRIC] Fixed serversided config values not showing up in config.
 - [FORGE] Updated Apoli Forge dependency to 2.6.1.1. (Equivalent Origins Forge: 1.7.1.1)